improve responsiveness of bookmarks
