--- shared/data/src/commonMain/kotlin/fr/androidmakers/store/graphql/ApolloClientBuilder.kt
@@ -1,7 +1,0 @@
-package fr.androidmakers.store.graphql
-
-import com.apollographql.apollo3.ApolloClient
-
-expect class ApolloClientBuilder {
-  fun build(): ApolloClient
-}
