+++ shared/di/src/androidMain/kotlin/fr/androidmakers/di/DataModule.android.kt
@@ -2,17 +2,16 @@ package fr.androidmakers.di
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
-import fr.androidmakers.store.graphql.ApolloClientBuilder
+import com.apollographql.apollo3.cache.normalized.sql.SqlNormalizedCacheFactory
+import fr.androidmakers.store.graphql.ApolloClient
 import fr.androidmakers.store.local.createDataStore
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 
 actual val dataPlatformModule = module {
   single {
-    ApolloClientBuilder(
-      androidContext(),
-      "https://androidmakers.fr/graphql",
-      "androidmakers2024",
+    ApolloClient(
+      SqlNormalizedCacheFactory(context = get()),
       get()
     )
   }
